requirements.txt
os
langchain
lanchain.core
langchain_openai
dotenv_python
numpy
